chore: some restructuring
